Finish 2.20*: Phonebook step12
Open your application in two browsers. **If you delete a person in
browser 1** a short while before attempting to _change the person's
phone number_ in browser 2, you will get the a error message.

Fix the issue according to the example shown in [promise and
errors](https://fullstackopen.com/en/part2/altering_data_in_server#promises-and-errors)
in part 2.

Modify the example so that the user is shown a message when
the operation does not succeed. The messages shown for successful and
unsuccessful events should look different.