Finish 3.5: Phonebook backend step5
Expand the backend so that new phonebook entries can be added by
making HTTP POST requests to the address http://localhost:3001/api/persons.

Generate a new id for the phonebook entry with the
[Math.random](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Math/random)
function. Use a big enough range for your random values so that the
likelihood of creating duplicate id's is small.