Finish 3.9 phonebook backend step9
Make the backend work with the frontend from the previous part.

You will probably have to do some small changes to the frontend, at
least to the URLs for the backend.

Remember to keep the developer console open in your browser.
If some HTTP requests fail, you should check from the _Network_-tab
what is going on.

Keep an eye on the backend's console as well. If you did not do
the previous exercise, it is worth it to print the request data
or _request.body_ to the console in the event handler responsible for
POST requests.