Finish 3.12: Command-line database
Create a cloud-based MongoDB database for the phonebook application
with MongoDB Atlas.

Create a _mongo.js_ file in the project directory, that can be used
for adding entries to the phonebook, and for listing all of the
existing entries in the phonebook.

The application should work as follows. You use the program by passing
three command-line arguments (the first is the password), e.g.:

1. `node mongo.js yourpassword Anna 040-1234556`
   Output: added Anna number 040-1234556 to phonebook
2. `node mongo.js yourpassword`
   Output: Then the program should display all of the entries in the phonebook:

   phonebook:
   Anna 040-1234556
   Arto Vihavainen 045-1232456
   Ada Lovelace 040-1231236