Finish 4.7*: helper functions and unit tests, step5
Define a function called _mostLikes_ that receives an array of blogs
as its parameter. The function returns the author, whose blog posts
have the largest amount of likes. The return value also contains the
total number of likes that the author has received.

If there are many top bloggers, then it is enough to show any one of them.