4.13 cont.: add the checking the blogs stored in the database function into a new file
Both tests check the state stored in the database after the saving
operation, by fetching all the blogs of the application.

```js
const response = await api.get('/api/blogs')
```

The same verification steps will repeat in other tests later on, and
it is a good idea to extract these steps into helper functions.

Add the function into a new file called _tests/test_helper.js_ that is
in the same directory as the test file.