Finish 4.14 Blog list expansions, step2
Implement functionality for updating the information of an individual
blog post.

Use async/await.

The application mostly needs to update the amount of _likes_ for a
blog post. You can implement this functionality the same way that we
implemented updating notes in [part
3](https://fullstackopen.com/en/part3/saving_data_to_mongo_db#other-operations).

Feel free to implement tests for the functionality if you want
to. Otherwise verify that the functionality works with Postman or some
other tool.