4.15 cont.: Create User model
Define the model for representing a user in the _models/user.js_ file

The ids of the blogs are stored within the user document as an array
of Mongo ids.

Expand the schema of the blog defined in the _model/blog.js_ file so
that the blog contains information about the user who created it.

In stark contrast to the conventions of relational databases,
_references are now stored in both documents_: the blog references the
user who created it, and the user has an array of references to all of
the blogs created by them.