4.15 cont.: Create users router for creating new users
Implement a route for creating new users. Users have
- a unique _username_
- a _name_
- and something called a _passwordHash_.

The password hash is the output of a [one-way hash
function](https://en.wikipedia.org/wiki/Cryptographic_hash_function)
applied to the user's password.

It is never wise to store unencrypted plain text passwords in the
database!

Install the [bcrypt](https://github.com/kelektiv/node.bcrypt.js)
package for generating the password hashes

Define a separate _router_ for dealing with users in a new
_controllers/users.js_ file.

Take the router into use in our application in the _app.js_ file, so
that it handles requests made to the _/api/users_ url.

The password sent in the request is _not_ stored in the database. We
store the _hash_ of the password that is generated with the
_bcrypt.hash_ function.