Finish 4.16*: bloglist expansion, step4
Password restrictions and test

Add a feature which adds the following restrictions to creating new
users:

- password must be given.
- password must be at least 3 characters long.

Do not test password restrictions with Mongoose validations. It is not
a good idea because the password received by the backend and the
password hash saved to the database are not the same thing. The
password length should be validated in the controller like we did in
[part 3](https://fullstackopen.com/en/part3/validation_and_es_lint)
before using Mongoose validation.