5.2 cont.: Make the login 'permanent' by using the local storage.
Solved by saving the login details to [local
storage](https://developer.mozilla.org/en-US/docs/Web/API/Storage). Local
Storage is a
[key-value](https://en.wikipedia.org/wiki/Key-value_database) database
in the browser.

A _value_ corresponding to a certain _key_ is saved to the database
with method
[setItem](https://developer.mozilla.org/en-US/docs/Web/API/Storage/setItem). For
example:

```js
window.localStorage.setItem('name', 'juha tauriainen')
```

saves the string given as the second parameter as the value of key _name_.

The value of a key can be found with method
[getItem](https://developer.mozilla.org/en-US/docs/Web/API/Storage/getItem):

```js
window.localStorage.getItem('name')
```

Values saved to the storage are
[DOMstrings](https://developer.mozilla.org/en-US/docs/Web/API/DOMString),
so we cannot save a JavaScript object as is. The object has to be
parsed to JSON first, with the method
_JSON.stringify_. Correspondingly, when a JSON object is read from the
local storage, it has to be parsed back to JavaScript with
_JSON.parse_.

We still have to modify our application so that when we enter the
page, the application checks if user details of a logged-in user can
already be found on the local storage. If they can, the details are
saved to the state of the application and to _noteService_.

The right way to do this is with an [effect
hook](https://reactjs.org/docs/hooks-effect.html): a mechanism we
first encountered in [part
2](https://fullstackopen.com/en/part2/getting_data_from_server#effect-hooks),
and used to fetch notes from the server.

The empty array as the parameter of the effect ensures that the effect
is executed only when the component is rendered [for the first
time](https://reactjs.org/docs/hooks-reference.html#conditionally-firing-an-effect).