5.3 cont.: save token to header
The token returned with a successful login is saved to the
application's state - the _user_'s field _token_.

Adding the token of the logged-in user to the Authorization header of
the HTTP request.

The event handler responsible for log in must be changed to call the
method `noteService.setToken(user.token)` with a successful log in.