Finish 5.7* Blog list frontend, step7
Let's add each blog a button, which controls if all of the details
about the blog are shown or not.

Full details of the blog open when the button is clicked.

And the details are hidden when the button is clicked again.

At this point the _like_ button does not need to do anything.