Finish 5.10*: Blog list frontend, step10
Add a new button for deleting blog posts. Also implement the logic for
deleting blog posts in the backend.

The confirmation dialog for deleting a blog post is easy to implement
with the `window.confirm` function.

Show the button for deleting a blog post only if the blog post was
added by the user.