Finish 6.1: unicafe revisited, step1
**Implement the reducer and its tests.**

In the tests, make sure that the reducer is an _immutable function_
with the _deep-freeze_-library.

Ensure that the provided first test passes, because Redux expects that
the reducer returns a sensible original state when it is called so
that the first parameter _state_, which represents the previous state,
is _undefined_.

Start by expanding the reducer so that both tests pass. Then add the
rest of the tests, and finally the functionality which they are
testing.