Finish 6.10 Better anecdotes, step8
The application has a ready-made body for the _Notification_
component.

Extend the component so that it renders the message stored in the
redux store, making the component to take the form:

```js
import React from 'react'
import { useSelector } from 'react-redux'
const Notification = () => {
  const notification = useSelector(/* something here */)  const style = {
    border: 'solid',
    padding: 10,
    borderWidth: 1
  }
  return (
    <div style={style}>
      {notification}    </div>
  )
}
```

You will have to make changes to the application's existing
reducer. Create a separate reducer for the new functionality and
refactor the application so that it uses a combined reducer as shown
in this part of the course material.

The application does not have to use the _Notification_ component in
any intelligent way at this point in the exercises. It is enough for
the application to display the initial value set for the message in
the _notificationReducer_.