Finish 6.11 Better anecdotes, step9
Extend the application so that it uses the _Notification_ component to
display a message for the duration of 5 seconds when the user votes
for an anecdote or creates a new anecdote.

It's recommended to create separate [action
creators](https://redux.js.org/basics/actions#action-creators) for
setting and removing notifications.