Finish 6.12* Better anecdotes, step10
Implement filtering for the anecdotes that are displayed to the user.

Store the state of the filter in the redux store. It is recommended to
create a new reducer and action creators for this purpose.

Create a new _Filter_ component for displaying the filter. You can use
the following code as a template for the component:

```js
import React from 'react'

const Filter = () => {
  const handleChange = (event) => {
    // input-field value is in variable event.target.value
  }
  const style = {
    marginBottom: 10
  }

  return (
    <div style={style}>
      filter <input onChange={handleChange} />
    </div>
  )
}

export default Filter
```