6.13 cont.: fetch data and initialize anecdotes
Create a method into the file _services/anecdotes.js_, which uses _axios_
to fetch data from the backend.

Add support in the reducer for the action _INIT_ANECDOTES_, using
which the initialization can be done by dispatching a single
action. Also create an action creator function _initializeAnecdotes_.

Initialize the state based on the data on the server is to fetch the
anecdotes in the file _index.js_ and dispatch to the action _NEW_ANECDOTES_.