Finish 6.13 Anecdotes and the backend, step1
When the application launches, fetch the anecdotes from the backend
implemented using json-server.

======

Move the initialization of the notes into the _App_ component, and, as
usual when fetching data from a server, we'll use the _effect hook_.

Using the useEffect hook causes an eslint-warning. We can get rid of
it by doing the following:

```js
const App = () => {
  const dispatch = useDispatch()
  useEffect(() => {
    noteService
      .getAll().then(notes => dispatch(initializeNotes(notes)))
  }, [dispatch])
  // ...
}
```

Now the variable _dispatch_ we define in the _App_ component, which practically is the dispatch function of the redux-store, has been added to the array useEffect receives as a parameter.
**If** the value of the dispatch-variable would change during runtime,
the effect would be executed again. This however cannot happen in our
application, so the warning is unnecessary.

Another way to get rid of the warning would be to disable eslint on that line:

```js
const App = () => {
  const dispatch = useDispatch()
  useEffect(() => {
    noteService
      .getAll().then(notes => dispatch(initializeNotes(notes)))
  },[]) // eslint-disable-line react-hooks/exhaustive-deps
  // ...
}
```

Generally disabling eslint when it throws a warning is not a good
idea. Even though the eslint rule in question has caused some
[arguments](https://github.com/facebook/create-react-app/issues/6880),
we will use the first solution.

More about the need to define the hooks dependencies in [the react
documentation](https://reactjs.org/docs/hooks-faq.html#is-it-safe-to-omit-functions-from-the-list-of-dependencies).