Finish 6.15 Anecdotes and the backend, step3
Modify the initialization of redux-store to happen using asynchronous
action creators, which are made possible by the _redux-thunk_-library.

==========

It is not great that the communication with the server happens inside
the functions of the components. It would be better if the
communication could be abstracted away from the components, such that
they don't have to do anything else but call the appropriate _action
creator_. As an example, _App_ would initialize the state of the
application as follows:

```js
const App = () => {
  const dispatch = useDispatch()

  useEffect(() => {
    dispatch(initializeNotes()))
  },[dispatch])

  // ...
}
```

Both components would only use the function provided to them as a prop
without caring about the communication with the server that is
happening in the background.

Now let's install the
[redux-thunk](https://github.com/gaearon/redux-thunk)-library, which
enables us to create _asynchronous actions_. Installation is done with
the command:

```language-js
npm install --save redux-thunk
```

The redux-thunk-library is a so-called _redux-middleware_, which must
be initialized along with the initialization of the store. While we're
here, let's extract the definition of the store into its own file
_src/store.js_.