6.19 cont.: get rid of useSelector
The _connect_ function can be used for transforming "regular" React
components so that the state of the Redux store can be "mapped" into
the component's props.

The component needs the list of anecdotes and the value of the filter from
the Redux store. The _connect_ function accepts a so-called
[mapStateToProps](https://github.com/reduxjs/react-redux/blob/master/docs/api/connect.md#mapstatetoprops-state-ownprops--object)
function as its first parameter. The function can be used for defining
the props of the _connected component_ that are based on the state of
the Redux store.

The _Anecdotes_ component can access the state of the store directly,
e.g. through _props.anecdotes_ that contains the list of anecdotes.
Similarly, _props.filter_ references the value of the filter.