Finish 6.19 anecdotes and connect, step1
The _redux store_ is currently passed to all of the components through props.

Add the [react-redux](https://github.com/reactjs/react-redux) package
to your application, and modify the _AnecdoteList_ so that it accesses
the store's state with the help of the _connect_ function.

===========

The second parameter of the _connect_ function can be used for
defining
[mapDispatchToProps](https://github.com/reduxjs/react-redux/blob/master/docs/api/connect.md#mapdispatchtoprops-object--dispatch-ownprops--object)
which is a group of _action creator_ functions passed to the connected
component as props.

There is no need to call the _dispatch_ function separately since
_connect_ has already modified the _voteById_ action creator into a
form that contains the dispatch.