Finish 6.21 anecdotes, the grand finale
You (probably) have one nasty bug in your application. If the user
clicks the vote button multiple times in a row, the notification is
displayed funnily. For example if a user votes twice in three seconds,
the last notification is only displayed for two seconds (assuming the
notification is normally shown for 5 seconds). This happens because
removing the first notification accidentally removes the second
notification.

Fix the bug so that after multiple votes in a row, the notification
for the last vote is displayed for five seconds.

This can be done by cancelling the removal of the previous
notification when a new notification is displayed whenever necessary.

The
[documentation](https://developer.mozilla.org/en-US/docs/Web/API/WindowOrWorkerGlobalScope/setTimeout)
for the setTimeout function might also be useful for this.