Finish 7.1: routed anecdotes, step1
Add React Router to the application so that by clicking links in the
_Menu_-component the view can be changed.

- At the root of the application, meaning the path _/_, show the list
  of anecdotes.
- The creation of a new anecdote should happen e.g. in the path
  _create_.

The _Footer_-component should always be visible at the bottom.

========

Note that the order of the components is important. If we would put
the _Home_-component, whose path is  _path="/"_, first, nothing else
would ever get rendered because the "nonexistent" path "/" is the
start of every path:

```js
<Switch>
  <Route path="/">
    <Home />
  </Route>
  <Route path="/notes">
    <Notes />
  </Route>
  // ...
</Switch>
```