Finish 7.2: routed anecdotes, step2
Implement a view for showing a single anecdote:

Navigating to the page showing the single anecdote is done by clicking
the name of that anecdote.

=======

We define the route rendering a specific note "express style" by
marking the parameter with a colon _:id_

```js
<Route path="/notes/:id">
```

When a browser navigates to the url for a specific note, for example
_/notes/3_, we render the _Note_ component:

```js
import {
  // ...
  useParams} from "react-router-dom"

const Note = ({ notes }) => {
  const id = useParams().id  const note = notes.find(n => n.id === Number(id))
  return (
    <div>
      <h2>{note.content}</h2>
      <div>{note.user}</div>
      <div><strong>{note.important ? 'important' : ''}</strong></div>
    </div>
  )
}
```

The _Note_ component receives all of the notes as props _notes_, and
it can access the url parameter (the id of the note to be displayed)
with the
[useParams](https://reacttraining.com/react-router/web/api/Hooks/useparams)
function of the react-router.