Finish 7.3: routed anecdotes, step3
The default functionality of the creation form is quite confusing,
because nothing seems to be happening after creating a new anecdote
using the form.

Improve the functionality such that after creating a new anecdote the
application transitions automatically to showing the view for all
anecdotes and the user is shown a notification informing them of this
successful creation for the next 10 seconds.