Finish 7.5: anecdotes and hooks step2
Add a button to the form that you can use to clear all the input
fields:

Expand the functionality of the _useField_ hook so that it offers a
new _reset_ operation for clearing the field.

Depending on your solution you may see the following warning in your
console:

```
Warning: Invalid value for prop `reset` on <input> tag. Either remove
it from the element, or pass a string or number value to keep it in
the DOM. For details, see https://fb.me/react-attribute-behavior
```

We will return to this warning in the next exercise.