Finish 7.7: country hook
The application can be used to search for country details from the
https://restcountries.eu/ interface. If country is found, the details
of the country are displayed.

If country is not found, message `not found` is displayed to the user.

The application is otherwise complete, but in this exercise you have
to implement a custom hook _useCountry_, which can be used to search
for the details of the country given to the hook as a parameter.

Use the api endpoint [full
name](https://restcountries.eu/#api-endpoints-full-name) to fetch
country details in a _useEffect_-hook within your custom hook.

Note, that in this exercise it is essential to use useEffect's [second
parameter](https://reactjs.org/docs/hooks-reference.html#conditionally-firing-an-effect)
array to control when the effect function is executed.