Finish 7.10: redux, step2 and 7.11: redux, step3
7.10: redux, step2

Store the information about blog posts in the Redux store. In this
exercise it is enough that you can see the blogs in backend and create
a new blog.

You are free to manage the state for logging in and creating new blog
posts by using the internal state of React components.

7.11: redux, step3

Expand your solution so that it is again possible to like and delete a
blog.