Finish 8.1: The number of books and authors
Implement queries _bookCount_ and _authorCount_ which return the
number of books and the number of authors.

The query

```gql
query {
  bookCount
  authorCount
}
```

should return

```json
{
  "data": {
    "bookCount": 7,
    "authorCount": 5
  }
}
```

==========

When Apollo-server is run on development mode (_node filename.js_), it
starts a
[GraphQL-playground](https://www.apollographql.com/docs/apollo-server/testing/graphql-playground/)
to address http://localhost:4000/graphql. This is very useful for a
developer, and can be used to make queries to the server.