Finish 8.4: Books of an author
Modify the _allBooks_ query so, that a user can give an optional
parameter _author_. The response should include only books written by
that author.

For example query

```gql
query {
  allBooks(author: "Robert Martin") {
    title
  }
}
```

should return

```json
{
  "data": {
    "allBooks": [
      {
        "title": "Clean Code"
      },
      {
        "title": "Agile software development"
      }
    ]
  }
}
```

========

The second parameter, _args_, contains the parameters of the query.
The resolver then returns from the array _books_ the book whose
author is the same as the value of _args.author_
The resolver does not need the first parameter _root_.

In fact all resolver functions are given [four
parameters](https://www.apollographql.com/docs/graphql-tools/resolvers.html#Resolver-function-signature). With
JavaScript the parameters don't have to be defined, if they are not
needed.