Finish 8.6: Adding a book
Implement mutation _addBook_, which can be used like this:

```gql
mutation {
  addBook(
    title: "NoSQL Distilled",
    author: "Martin Fowler",
    published: 2012,
    genres: ["database", "nosql"]
  ) {
    title,
    author
  }
}
```

The mutation works even if the author is not already saved to the
server:

```gql
mutation {
  addBook(
    title: "Pimeyden tango",
    author: "Reijo Mäki",
    published: 1997,
    genres: ["crime"]
  ) {
    title,
    author
  }
}
```

If the author is not yet saved to the server, a new author is added to
the system. The birth years of authors are not saved to the server
yet, so the query

```gql
query {
  allAuthors {
    name
    born
    bookCount
  }
}
```

returns

```json
{
  "data": {
    "allAuthors": [
      // ...
      {
        "name": "Reijo Mäki",
        "born": null,
        "bookCount": 1
      }
    ]
  }
}
```

=========

Adding a mutation requires both schema and resolver.

The schema is given the details of method parameters.

The schema also has a return value. The return value is type Person,
the idea being that the details of the added person are returned if
the operation is successful and if not, null.

Value for the field id is not given as a parameter. Generating an id
is better left for the server.