8.10 cont.: implement functionality for adding new books
The hook-function
[useMutation](https://www.apollographql.com/docs/react/v3.0-beta/api/react/hooks/#usemutation)
provides the functionality for making mutations.

We can define mutation function using the _useMutation_-hook.

The hook returns an _array_, first element of which contains the
result of the mutation.

```js
const [ createPerson ] = useMutation(CREATE_PERSON)
```

The query variables receive values when the query is made:

```js
createPerson({  variables: { name, phone, street, city } })
```

New persons are added just fine, but the screen is not updated. The
reason being that Apollo Client cannot automatically update the cache
of an application, so it still contains the state from before the
mutation.

We could update the screen by reloading the page, as the cache is
emptied when the page is reloaded. However there must be a better way
to do this.