Finish 8.10: Adding a book
Implement a possibility to add new books to your application.

Make sure that the Authors and Books views are kept up to date after a
new book is added.

In case of problems when making queries or mutations, check from
developer console what the server response is.

=========

**Updating the cache**

There are few different solutions for this.

1. One way is to make the query for all persons
[poll](https://www.apollographql.com/docs/react/v3.0-beta/data/queries/#polling)
the server, or make the query repeatedly.

The change is small. Let's set the query to poll every two seconds:

```js
  const result = useQuery(ALL_PERSONS, {
    pollInterval: 2000
  })
```

The solution is simple, and every time a user adds a new person, it
appears immediately on the screens of all users.

The bad side of the solution is all the pointless web traffic.

2. Another easy way to keep the cache in sync is to use the
_useMutation_-hook's
[refetchQueries](https://www.apollographql.com/docs/react/v3.0-beta/api/react/hooks/#params-2)
parameter to define, that the query fetching all persons is done again
whenever a new person is created.

```js
  const [ createPerson ] = useMutation(CREATE_PERSON, {
    refetchQueries: [ { query: ALL_PERSONS } ]  })
```

The pros and cons of this solution are almost opposite of the previous
one. There is no extra web traffic, because queries are not done just
in case.  However if one user now updates the state of the server, the
changes do not show to other users immediately.

There are other ways to update the cache. More about those later in
this part.