8.13 cont.: install mongoose and config connection
Change the library application so that it saves the data to a
database. You can find the _mongoose schema_ for books and authors
from
[here](https://github.com/fullstack-hy2020/misc/blob/master/library-schema.md).

Change the book graphql schema a little

```gql
type Book {
  title: String!
  published: Int!
  author: Author!
  genres: [String!]!
  id: ID!
}
```

so that instead of just the author's name, the book object contains
all the details of the author.

========

Book query should change to

```gql
mutation {
  addBook(
    title: "NoSQL Distilled",
    author: "Martin Fowler",
    published: 2012,
    genres: ["database", "nosql"]
  ) {
    title,
    author {  // <--- Here changed
      name
    }
  }
}
```

========

Install mongoose and mongoose-unique-validator:

```language-js
npm install mongoose mongoose-unique-validator --save
```

We will imitate what we did in parts
[3](https://fullstackopen.com/en/part3/saving_data_to_mongo_db) and
[4](https://fullstackopen.com/en/part4/structure_of_backend_application_introduction_to_testing).