Finish 8.13: Database, part 1
You can assume that the user will not try to add faulty books or authors, so you don't have to care about validation errors.

The following things do _not_ have to work just yet

* _allBooks_ query with parameters
* _bookCount_ field of an author object
* _author_ field of a book
* _editAuthor_ mutation

==========

The changes are pretty straightforward. However there are a few
noteworthy things. As we remember, in Mongo the identifying field of
an object is called __id_ and we previously had to parse the name of
the field to _id_ ourselves. Now GraphQL can do this automatically.

Another noteworthy thing is that the resolver functions now return a
_promise_, when they previously returned normal objects. When a
resolver returns a promise, Apollo server [sends
back](https://www.apollographql.com/docs/apollo-server/data/data/#resolver-results)
the value which the promise resolves to.

For example if the following resolver function is executed,

```js
allPersons: (root, args) => {
  return Person.find({})
},
```

Apollo server waits for the promise to resolve, and returns the
result. So Apollo works roughly like this:

```js
Person.find({}).then( result => {
  // return the result
})
```