8.16 cont.: create resolvers for `createUser`, `login` and `me`
Create resolvers for query _me_ and the new mutations _createUser_ and
_login_. Like in the course material, you can assume all users have
the same hardcoded password.

======

The new user mutation is straightforward. The log in mutation checks
if the username/password pair is valid. And if it is indeed valid, it
returns a jwt-token familiar from part 4.

Just like in the previous case with REST, the idea now is that a
logged in user adds a token they receive upon log in to all of their
requests. And just like with REST, the token is added to GraphQL
queries using the Authorization header.

In the GraphQL-playground the header is added to a query in `HTTP
HEADERS` like so

```
{
  "Authorization": "bearer xxxxxxxxxxx"
}
```

===========

Let's now expand the definition of the _server_ object by adding a
third parameter
[context](https://www.apollographql.com/docs/apollo-server/data/data/#context-argument)
to the constructor call:

```js
const server = new ApolloServer({
  typeDefs,
  resolvers,
  context: async ({ req }) => {
    const auth = req ? req.headers.authorization : null
    if (auth && auth.toLowerCase().startsWith('bearer ')) {
      const decodedToken = jwt.verify(
        auth.substring(7), JWT_SECRET
      )
      const currentUser = await User.findById(decodedToken.id).populate('friends')
      return { currentUser }
    }
  }
})
```

The object returned by context is given to all resolvers as their
_third parameter_. Context is the right place to do things which are
shared by multiple resolvers, like [user
identification](https://blog.apollographql.com/authorization-in-graphql-452b1c402a9).

So our code sets the object corresponding to the user who made the
request to the _currentUser_ field of the context. If there is no user
connected to the request, the value of the field is undefined.

================

The resolver of the _me_ query is very simple, it just returns the
logged in user it receives in the _currentUser_ field of the third
parameter of the resolver, _context_. It's worth noting that if there
is no logged in user, i.e there is no valid token in the header
attached to the request, the query returns _null_:

```js
Query: {
  // ...
  me: (root, args, context) => {
    return context.currentUser
  }
},
```