8.18 cont.: Implement the funtion to login
Add variable _token_ to the application's state. It will contain
user's token when a is logged in. If _token_ is undefined, we render
the _LoginForm_-component responsible for user login. The component
receives an error handler and the _setToken_-function as parameters.

Next we define a mutation for logging in

```js
export const LOGIN = gql`
  mutation login($username: String!, $password: String!) {
    login(username: $username, password: $password)  {
      value
    }
  }
`
```

The _LoginForm_-component works pretty much just like all other
components doing mutations we have previously created.

We are using an effect hook again. Here it's used to save the token's
value to the state of the _App_ component and the local storage after
the server has responded to the mutation.

Use of the effect hook is necessary to avoid an endless rendering
loop.